backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==41.0.7
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
fernet==1.0.1
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
jose==1.0.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.0
openai==1.79.0
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pip-tools==7.4.1
playwright==1.52.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyaes==1.6.1
pyasn1==0.4.7
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
PyPDF2==3.0.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
